+++ app/src/main/java/com/catpuppyapp/puppygit/git/PatchFile.kt
@@ -1,0 +1,13 @@
+package com.catpuppyapp.puppygit.git
+
+data class PatchFile(
+    /**
+     * output file path
+     */
+    val outFileFullPath:String? = null,
+
+    /**
+     * diff content (patch content)
+     */
+    val content:String? = null,
+)